+++ src/workflows-app/RunDetails.js
@@ -4,7 +4,6 @@ import { div, h, span } from 'react-hyperscript-helpers';
 import { Link } from 'src/components/common';
 import { centeredSpinner, icon } from 'src/components/icons';
 import { collapseStatus } from 'src/components/job-common';
-import { LogViewer } from 'src/components/UriViewer/LogViewer';
 import { Ajax } from 'src/libs/ajax';
 import { notify } from 'src/libs/notifications';
 import { useCancellation, useOnMount, usePollingEffect } from 'src/libs/react-utils';
@@ -14,6 +13,7 @@ import { cond, newTabLinkProps } from 'src/libs/utils';
 import CallTable from 'src/pages/workspaces/workspace/jobHistory/CallTable';
 import InputOutputModal from 'src/workflows-app/components/InputOutputModal';
 import { HeaderSection, statusType, SubmitNewWorkflowButton } from 'src/workflows-app/components/job-common';
+import { LogViewer } from 'src/workflows-app/components/LogViewer';
 import { TroubleshootingBox } from 'src/workflows-app/components/TroubleshootingBox';
 import { WorkflowInfoBox } from 'src/workflows-app/components/WorkflowInfoBox';
 import { doesAppProxyUrlExist, loadAppUrls } from 'src/workflows-app/utils/app-utils';
